papermind_backend/requirements.txt

djangorestframework
pdfminer.six
gemini
langchain
pymupdf
numpy
pandas
spacy
transformers
sentence-transformers
faiss-cpu